Move expensive logic from new to next_block, which is called less often
